kafka/kafka-connect/requirements.txt
git+https://github.com/openai/whisper.git
boto3
pydub
python-dotenv
ffmpeg-python
kafka-python
